Refactor location package into world package
