+++ src/main/java/me/darkeyedragon/randomtp/RandomTeleport.java
@@ -8,13 +8,13 @@ import me.darkeyedragon.randomtp.config.ConfigHandler;
 import me.darkeyedragon.randomtp.eco.EcoHandler;
 import me.darkeyedragon.randomtp.listener.PluginLoadListener;
 import me.darkeyedragon.randomtp.listener.WorldLoadListener;
-import me.darkeyedragon.randomtp.location.LocationFactory;
-import me.darkeyedragon.randomtp.location.LocationSearcher;
 import me.darkeyedragon.randomtp.validator.ChunkValidator;
 import me.darkeyedragon.randomtp.validator.ValidatorFactory;
 import me.darkeyedragon.randomtp.world.LocationQueue;
 import me.darkeyedragon.randomtp.world.QueueListener;
 import me.darkeyedragon.randomtp.world.WorldQueue;
+import me.darkeyedragon.randomtp.world.location.LocationFactory;
+import me.darkeyedragon.randomtp.world.location.LocationSearcher;
 import net.milkbowl.vault.economy.Economy;
 import org.bukkit.Bukkit;
 import org.bukkit.ChatColor;
