+++ src/main/java/me/darkeyedragon/randomtp/command/TeleportCommand.java
@@ -11,11 +11,11 @@ import me.darkeyedragon.randomtp.config.data.ConfigMessage;
 import me.darkeyedragon.randomtp.config.data.ConfigQueue;
 import me.darkeyedragon.randomtp.config.data.ConfigWorld;
 import me.darkeyedragon.randomtp.eco.EcoHandler;
-import me.darkeyedragon.randomtp.location.LocationFactory;
-import me.darkeyedragon.randomtp.location.WorldConfigSection;
 import me.darkeyedragon.randomtp.world.LocationQueue;
 import me.darkeyedragon.randomtp.world.QueueListener;
 import me.darkeyedragon.randomtp.world.WorldQueue;
+import me.darkeyedragon.randomtp.world.location.LocationFactory;
+import me.darkeyedragon.randomtp.world.location.WorldConfigSection;
 import org.bukkit.ChatColor;
 import org.bukkit.Location;
 import org.bukkit.Particle;
