+++ src/main/java/me/darkeyedragon/randomtp/config/ConfigHandler.java
@@ -2,8 +2,8 @@ package me.darkeyedragon.randomtp.config;
 
 import me.darkeyedragon.randomtp.RandomTeleport;
 import me.darkeyedragon.randomtp.config.data.*;
-import me.darkeyedragon.randomtp.location.WorldConfigSection;
 import me.darkeyedragon.randomtp.util.TimeUtil;
+import me.darkeyedragon.randomtp.world.location.WorldConfigSection;
 import org.bukkit.Bukkit;
 import org.bukkit.ChatColor;
 import org.bukkit.World;
